+++ .gitignore
@@ -1,0 +1,7 @@
+.venv/
+venv
+*.pyc
+__pycache__
+dist/
+*.log
+result/
