+++ __init__.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python
+# coding=utf-8
+
+'''
+Version: 0.1
+Autor: zmf96
+Email: zmf96@qq.com
+Date: 2022-03-04 07:26:01
+LastEditors: zmf96
+LastEditTime: 2022-03-04 07:26:01
+FilePath: /__init__.py
+Description: 
+'''
