+++ emailall/modules/search/ask.py
@@ -7,55 +7,53 @@
 
 
 import time
-from common.search import Search
-from config.setting import emails
+from emailall.common.search import Search
+from emailall.config.log import logger
 from lxml import etree
+from emailall.config import settings
 
 
-class Baidu(Search):
+class Ask(Search):
+
     def __init__(self, domain):
         Search.__init__(self)
         self.domain = domain
         self.module = 'Search'
-        self.source = 'BaiduSearch'
-        self.addr = 'https://www.baidu.com/s'
+        self.source = 'AskSearch'
+        self.addr = 'https://www.ask.com/web'
         self.urls = list()
-        self.limit_num = 500
+        self.timeout = 15
+        self.limit_num = 50
         self.per_page_num = 10
 
-    def get_url(self, html):
-        data = []
+    def get_urls(self, html):
         html = etree.HTML(html)
-        urls = html.xpath("//div[@id='content_left']//h3/a/@href")
+        urls = html.xpath("//div[@class='PartialSearchResults-results']//a/@href")
         for url in urls:
-            locat_url = self.match_location(url)
-            time.sleep(2)
-            data.append(locat_url)
-            self.urls.append(locat_url)
-        return data
+            self.urls.append(url)
 
     def search(self):
+        self.page_num = 1
         while True:
             time.sleep(self.delay)
+            self.proxy = settings.proxy
             query = 'intext:@' + self.domain
-            params = {'wd': query,
-                      'pn': self.page_num,
-                      'rn': self.per_page_num}
-            resp = self.get(self.addr, params)
-            if not resp:
-                return
-            data = self.get_url(resp.text)
-            for url in data:
+            params = {'q': query, 'page': self.page_num}
+            resp = self.get(self.addr, params=params)
+            if not resp and not hasattr(resp, 'text'):
+                logger.log('ERROR', f'For module {self.source}, you need to configure the proxy in setting.py file')
+                break
+            self.get_urls(resp.text)
+            self.proxy = None
+            for url in self.urls:
                 rep = self.get(url)
                 emails = self.match_emails(rep)
                 if emails:
                     self.results.update(emails)
                 else:
                     continue
-            self.page_num += self.per_page_num
-            if f'&amp;pn={self.page_num}&' not in resp.text:
-                break
-            if self.page_num >= self.limit_num:  # 搜索条数限制
+            self.page_num += 1
+            if '>Next<' not in resp.text:
                 break
 
     def run(self):
@@ -65,16 +63,14 @@ class Baidu(Search):
         self.save_json()
         self.save_res()
 
-
 def run(domain):
     """
     类统一调用入口
 
     :param str domain: 域名
     """
-    search = Baidu(domain)
+    search = Ask(domain)
     search.run()
-
 
 if __name__ == '__main__':
     run('example.com')
