+++ emailall/modules/search/baidu.py
@@ -7,53 +7,55 @@
 
 
 import time
-from common.search import Search
-from config.log import logger
+from emailall.common.search import Search
+from emailall.config.setting import emails
 from lxml import etree
-from config import settings
 
 
-class Ask(Search):
-
+class Baidu(Search):
     def __init__(self, domain):
         Search.__init__(self)
         self.domain = domain
         self.module = 'Search'
-        self.source = 'AskSearch'
-        self.addr = 'https://www.ask.com/web'
+        self.source = 'BaiduSearch'
+        self.addr = 'https://www.baidu.com/s'
         self.urls = list()
-        self.timeout = 15
-        self.limit_num = 50
+        self.limit_num = 500
         self.per_page_num = 10
 
-    def get_urls(self, html):
+    def get_url(self, html):
+        data = []
         html = etree.HTML(html)
-        urls = html.xpath("//div[@class='PartialSearchResults-results']//a/@href")
+        urls = html.xpath("//div[@id='content_left']//h3/a/@href")
         for url in urls:
-            self.urls.append(url)
+            locat_url = self.match_location(url)
+            time.sleep(2)
+            data.append(locat_url)
+            self.urls.append(locat_url)
+        return data
 
     def search(self):
-        self.page_num = 1
         while True:
             time.sleep(self.delay)
-            self.proxy = settings.proxy
             query = 'intext:@' + self.domain
-            params = {'q': query, 'page': self.page_num}
-            resp = self.get(self.addr, params=params)
-            if not resp and not hasattr(resp, 'text'):
-                logger.log('ERROR', f'For module {self.source}, you need to configure the proxy in setting.py file')
-                break
-            self.get_urls(resp.text)
-            self.proxy = None
-            for url in self.urls:
+            params = {'wd': query,
+                      'pn': self.page_num,
+                      'rn': self.per_page_num}
+            resp = self.get(self.addr, params)
+            if not resp:
+                return
+            data = self.get_url(resp.text)
+            for url in data:
                 rep = self.get(url)
                 emails = self.match_emails(rep)
                 if emails:
                     self.results.update(emails)
                 else:
                     continue
-            self.page_num += 1
-            if '>Next<' not in resp.text:
+            self.page_num += self.per_page_num
+            if f'&amp;pn={self.page_num}&' not in resp.text:
+                break
+            if self.page_num >= self.limit_num:  # 搜索条数限制
                 break
 
     def run(self):
@@ -63,14 +65,16 @@ class Ask(Search):
         self.save_json()
         self.save_res()
 
+
 def run(domain):
     """
     类统一调用入口
 
     :param str domain: 域名
     """
-    search = Ask(domain)
+    search = Baidu(domain)
     search.run()
+
 
 if __name__ == '__main__':
     run('example.com')
