+++ emailall/modules/search/sogou.py
@@ -5,8 +5,8 @@
 # @url    : http://github.com/Taonn
 # @author : Tao. (Taonn)
 import time
-from common.search import Search
-from config.log import logger
+from emailall.common.search import Search
+from emailall.config.log import logger
 from lxml import etree
 import re
